Addressed first code review issues and renamed extension
This addresses most of the issues brought up on the first code review,
and changes the name to Settings Guard for Firefox to better align with
naming policies. Also added support for claro-search as a homepage.

Here's what's pending at the moment:
* Changing the "All is well" alert to a separate dialog. This will also
fix the issue on Mac OS X where it appears before the main window.
* UX review for all UI.